Refactor script and fix json parsing Signed off by : Adam Pilkington apilkington@uk.ibm.com
