+++ interactivemap-app/src/main/java/net/wasdev/gameon/map/models/Coordinates.java
@@ -16,7 +16,12 @@
 package net.wasdev.gameon.map.models;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
 
+@JsonInclude(Include.NON_EMPTY)
+@JsonIgnoreProperties(ignoreUnknown=true)
 public class Coordinates {
 
     private int x;
