+++ interactivemap-app/src/main/java/net/wasdev/gameon/map/models/Exit.java
@@ -15,6 +15,7 @@
  *******************************************************************************/
 package net.wasdev.gameon.map.models;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -26,6 +27,7 @@ import io.swagger.annotations.ApiModelProperty;
         description = "An exit provides the door and URL required for the player to traverse the path "
                 + "to the next room.")
 @JsonInclude(Include.NON_EMPTY)
+@JsonIgnoreProperties(ignoreUnknown=true)
 public class Exit {
 
     /** Room id */
