+++ interactivemap-app/src/main/webapp/index.html
@@ -19,7 +19,7 @@
     <link rel="stylesheet" href="https://api.tiles.mapbox.com/mapbox.js/plugins/leaflet-locatecontrol/v0.43.0/L.Control.Locate.css">
     <link rel="stylesheet" href="assets/leaflet-groupedlayercontrol/leaflet.groupedlayercontrol.css">
     <link rel="stylesheet" href="assets/css/app.css">
-    
+
     <link rel="apple-touch-icon" sizes="76x76" href="https://game-on.org/icons/apple-touch-icon-76x76.png">
     <link rel="apple-touch-icon" sizes="120x120" href="https://game-on.org/icons/apple-touch-icon-120x120.png">
     <link rel="apple-touch-icon" sizes="152x152" href="https://game-on.org/icons/apple-touch-icon-152x152.png">
@@ -178,7 +178,7 @@
               <li class="list-group-item">Starting room : <img width="16" height="18" src="img/room-start.png"/></li>
               <li class="list-group-item">Healthy room : <img width="16" height="18" src="img/room-healthy.png"/></li>
               <li class="list-group-item">Sick room : <img width="16" height="18" src="img/room-sick.png"/></li>
-            </ul> 
+            </ul>
           </div>
           <div class="modal-footer">
             <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
@@ -286,7 +286,7 @@
            	  </td>
           	</tr>
           </table>
-          </div> <!-- radio div -->         
+          </div> <!-- radio div -->
           </div> <!-- modal-body -->
           <div class="modal-footer">
             <button type="button" id="wstestConnect-btn" class="btn btn-default" data-dismiss="modal">Send</button>
@@ -325,7 +325,7 @@
           	    </tr>
           	    <tr>
            	     <th>Full Name</th>
-           	     <td><input type="text" id="roomInfo_fullName" class="form-control" 
+           	     <td><input type="text" id="roomInfo_fullName" class="form-control"
            	     placeholder="Full name for your room"/></td>
            	   </tr>
            	   <tr>
@@ -341,32 +341,32 @@
             <table class='table table-striped table-bordered table-condensed'>
               <tr>
                 <th>Up</th>
-                <td><input type="text" id="roomInfo_u" class="form-control" 
+                <td><input type="text" id="roomInfo_u" class="form-control"
                 placeholder=""/></td>
               </tr>
               <tr>
                 <th>Down</th>
-                <td><input type="text" id="roomInfo_d" class="form-control" 
+                <td><input type="text" id="roomInfo_d" class="form-control"
                 placeholder=""/></td>
               </tr>
               <tr>
                 <th>North</th>
-                <td><input type="text" id="roomInfo_n" class="form-control" 
+                <td><input type="text" id="roomInfo_n" class="form-control"
                 placeholder=""/></td>
               </tr>
               <tr>
                 <th>East</th>
-                <td><input type="text" id="roomInfo_e" class="form-control" 
+                <td><input type="text" id="roomInfo_e" class="form-control"
                 placeholder=""/></td>
               </tr>
               <tr>
                 <th>South</th>
-                <td><input type="text" id="roomInfo_s" class="form-control" 
+                <td><input type="text" id="roomInfo_s" class="form-control"
                 placeholder=""/></td>
               </tr>
               <tr>
                 <th>West</th>
-                <td><input type="text" id="roomInfo_w" class="form-control" 
+                <td><input type="text" id="roomInfo_w" class="form-control"
                 placeholder=""/></td>
               </tr>
             </table>
@@ -383,7 +383,7 @@
 			    </tr>
 				<tr>
                 	<th>Target</th>
-                	<td><input type="text" id="roomInfo_target" class="form-control" 
+                	<td><input type="text" id="roomInfo_target" class="form-control"
                 	placeholder="The websocket end point"/></td>
               	</tr>
     		  </table>
@@ -426,7 +426,12 @@
     <script src="https://api.tiles.mapbox.com/mapbox.js/plugins/leaflet-locatecontrol/v0.43.0/L.Control.Locate.min.js"></script>
     <script src="assets/leaflet-groupedlayercontrol/leaflet.groupedlayercontrol.js"></script>
     <script src="https://kjur.github.io/jsrsasign/jsrsasign-latest-all-min.js"></script>
+    <script src="js/mapping.js"></script>
+    <script src="js/roomui.js"></script>
+    <script src="assets/js/poi.js"></script>
+    <script src="assets/js/search.js"></script>
     <script src="assets/js/app.js"></script>
+    <script src="assets/js/map.js"></script>
     <script src="js/sites.js"></script>
     <script src="js/wstest.js"></script>
     <script src="js/player.js"></script>
